requirements.txt
Pillow==10.4.0
requests==2.32.4
discord-webhook==1.3.0
python-dotenv==1.0.1